refactor loader to its own folder
