+++ src/helpers/tests/loadJson.test.ts
@@ -1,4 +1,4 @@
-import { loadJson } from '../loadJson';
+import { loadJson } from '../loaders/loadJson';
 import featuresReducer from '../../features/geoFeatures/geoFeaturesSlice';
 import { Feature, Geometry, GeoJsonProperties, FeatureCollection, LineString } from "geojson";
 import { createStore } from '@reduxjs/toolkit';
